Refactoring message cell and message content view.
